# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: MindAtlas API
  description: AI-powered mapping and organization web application API
  version: 0.1.0
  contact:
    name: MindAtlas Team
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token obtained from POST /api/auth/login.
        Tokens expire after 15 minutes; use /api/auth/refresh to renew.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        API key created via POST /api/keys.
        Used for programmatic/webhook access.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid email or password
      required: [error]
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        content:
          type: string
        content_type:
          type: string
          enum: [plain_text, link, code_snippet, note, task, idea, file, custom]
        metadata:
          type: object
          nullable: true
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
    ItemCreate:
      type: object
      properties:
        content:
          type: string
          description: The content of the item
        content_type:
          type: string
          enum: [plain_text, link, code_snippet, note, task, idea, file, custom]
        metadata:
          type: object
          nullable: true
      required: [content, content_type]
...
paths:
...
  /api/items:
    get:
      tags: [items]
      summary: List user items
      description: Returns all items owned by the authenticated user
      operationId: listItems
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  total:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [items]
      summary: Create an item
      description: >
        Creates a new item. Rate limited to 100 requests/min per user.
      operationId: createItem
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/integrations/notion/connect:
    post:
      tags: [integrations]
      summary: Connect Notion workspace
      description: Stores a Notion access token and workspace info
      operationId: notionConnect
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                access_token:
                  type: string
                workspace_id:
                  type: string
                workspace_name:
                  type: string
              required: [access_token, workspace_id, workspace_name]
      responses:
        '201':
          description: Notion connected
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace_id:
                    type: string
                  workspace_name:
                    type: string
                  connected_at:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/integrations/notion/import:
    post:
      tags: [integrations]
      summary: Import from Notion
      description: Imports pages from the connected Notion workspace as items
      operationId: notionImport
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                database_id:
                  type: string
                  description: Optional Notion database to import from
      responses:
        '200':
          description: Import results
          content:
            application/json:
              schema:
                type: object
                properties:
                  items_imported:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No Notion connection found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/integrations/notion/export:
    post:
      tags: [integrations]
      summary: Export to Notion
      description: Exports items to the connected Notion workspace
      operationId: notionExport
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                item_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Export results
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages_created:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No Notion connection found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/integrations/notion/status:
    get:
      tags: [integrations]
      summary: Get Notion connection status
      operationId: notionStatus
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Connection details
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected:
                    type: boolean
                  workspace_id:
                    type: string
                  workspace_name:
                    type: string
                  connected_at:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/integrations/notion/disconnect:
    delete:
      tags: [integrations]
      summary: Disconnect Notion workspace
      operationId: notionDisconnect
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Disconnected
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No active connection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...